Added simple caching mechanism. This is how it works: Until parsing stage, method calls are tested how many times they occurs, then if method occurences only once, it's result won't be cached, if it's called at least twice, then first method call is generated with caching result instruction, any further calls won't call method but restore previously stored score.
It's not perfect and doesn't cover more complex cases, the biggest problem I can't get manage to solve is how to determine if function should be cached if it has many seemingly different arguments. For example A(3 + 2) is basically the same A(2 + 3) and should be cached also.

I think It would be usefull to build some kind of arguments evaluator, maybe will try to cover this later. Solved #3